# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AIVEC Conference API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages/{key}:
    get:
      operationId: getPage
      tags: [pages]
      parameters:
        - $ref: "#/components/parameters/PageKey"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    put:
      operationId: updatePage
      tags: [pages]
      parameters:
        - $ref: "#/components/parameters/PageKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsPost"
    post:
      operationId: createNews
      tags: [news]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsPost"
  /news/{slug}:
    get:
      operationId: getNews
      tags: [news]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsPost"
  /news/id/{id}:
    patch:
      operationId: updateNews
      tags: [news]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsPost"
    delete:
      operationId: deleteNews
      tags: [news]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /sponsors:
    get:
      operationId: listSponsors
      tags: [sponsors]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsor"
    post:
      operationId: createSponsor
      tags: [sponsors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}:
    patch:
      operationId: updateSponsor
      tags: [sponsors]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
    delete:
      operationId: deleteSponsor
      tags: [sponsors]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
    PageKey:
      in: path
      name: key
      required: true
      schema:
        type: string
  schemas:
...
    AppUser:
      type: object
      required: [id, clerkId, email, role, createdAt]
      properties:
        id: { type: integer }
        clerkId: { type: string }
        email: { type: string }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        role: { type: string, enum: [user, admin] }
        createdAt: { type: string, format: date-time }
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [user, admin] }
...
    Page:
      type: object
      required: [id, key]
      properties:
        id: { type: integer }
        key: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        subtitleEn: { type: string }
        subtitleAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        sections: {}
        updatedAt: { type: string, format: date-time }
    PageInput:
      type: object
      properties:
        titleEn: { type: string }
        titleAr: { type: string }
        subtitleEn: { type: string }
        subtitleAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        sections: {}
...
    NewsPost:
      type: object
      required: [id, slug, titleEn, titleAr, published]
      properties:
        id: { type: integer }
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        excerptEn: { type: string }
        excerptAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        category: { type: string }
        coverUrl: { type: ["string", "null"] }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NewsInput:
      type: object
      required: [slug, titleEn, titleAr]
      properties:
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        excerptEn: { type: string }
        excerptAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        category: { type: string }
        coverUrl: { type: ["string", "null"] }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
    NewsUpdate:
      type: object
      properties:
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        excerptEn: { type: string }
        excerptAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        category: { type: string }
        coverUrl: { type: ["string", "null"] }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
...
    Sponsor:
      type: object
      required: [id, nameEn, nameAr, tier, order]
      properties:
        id: { type: integer }
        nameEn: { type: string }
        nameAr: { type: string }
        tier: { type: string, enum: [platinum, gold, silver, supporter, government] }
        logoUrl: { type: ["string", "null"] }
        websiteUrl: { type: ["string", "null"] }
        order: { type: integer }
    SponsorInput:
      type: object
      required: [nameEn, nameAr, tier]
      properties:
        nameEn: { type: string }
        nameAr: { type: string }
        tier: { type: string, enum: [platinum, gold, silver, supporter, government] }
        logoUrl: { type: ["string", "null"] }
        websiteUrl: { type: ["string", "null"] }
        order: { type: integer }
    SponsorUpdate:
      type: object
      properties:
        nameEn: { type: string }
        nameAr: { type: string }
        tier: { type: string, enum: [platinum, gold, silver, supporter, government] }
        logoUrl: { type: ["string", "null"] }
        websiteUrl: { type: ["string", "null"] }
        order: { type: integer }
...