# lib/api-zod/src/generated/types/sponsor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIVEC Conference API
 * OpenAPI spec version: 0.1.0
 */
import type { SponsorTier } from './sponsorTier';

export interface Sponsor {
  id: number;
  nameEn: string;
  nameAr: string;
  tier: SponsorTier;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  order: number;
}